requirements.txt
# Python >=3.8
# torch/torch-geometric/torch-sparse/torch-scatter/torch-cluster/torch-spline-conv
# 推荐用官方指令安装：https://pytorch-geometric.readthedocs.io/en/latest/install/installation.html

torch==2.1.0
torch-geometric==2.4.0
torch-sparse==0.6.17
torch-scatter==2.1.2
torch-cluster==1.6.3
torch-spline-conv==1.2.2
ogb==1.3.6
pyyaml==6.0.1
wandb==0.16.6
pandas==2.2.2
matplotlib==3.8.4
scipy==1.13.0
numpy==1.26.4
tqdm==4.66.4
